+++ src/Food/Food.cpp
@@ -3,7 +3,7 @@
 Food::Food(double x, double y, double width, double height) : Item(x, y, width, height, TYPE_FOOD){
 }
 
-void Food::tick(int delay){
+void Food::tick(double delay){
 	double y = this->getY();
 	if (!(this->isAtBottom)){
 		y = y + 1;
